requirements.txt
flask
requests
beautifulsoup4
pyjwt
pymongo
